requirements.txt
flask
SQLAlchemy
mysqlclient
py-bcrypt
Flask-JWT
requests
